fix(db): change primary key type from Integer to BigInteger for olymp and teacher verification models

+++ src/db/models/olymp.py
@@ -1,4 +1,4 @@
-from sqlalchemy import String, Integer, Date
+from sqlalchemy import String, BigInteger, Date
 from sqlalchemy.orm import Mapped, mapped_column
 
 from .base import BaseModel
@@ -7,7 +7,7 @@ from .base import BaseModel
 class OlympModel(BaseModel):
     __tablename__ = 'olymp'
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
     form: Mapped[str] = mapped_column(String(5), nullable=False)
     student_name: Mapped[str] = mapped_column(String(50), nullable=False)
     teacher_name: Mapped[str] = mapped_column(String(50), nullable=False)
